+++ src/com/seafile/seadroid2/data/ProgressMonitor.java
@@ -1,0 +1,7 @@
+package com.seafile.seadroid2.data;
+
+public interface ProgressMonitor {
+    void onProgressNotify(long total);
+
+    boolean isCancelled();
+}
