+++ src/com/seafile/seadroid2/ui/FetchFileDialog.java
@@ -14,8 +14,8 @@ import android.widget.TextView;
 import com.seafile.seadroid2.BrowserActivity;
 import com.seafile.seadroid2.R;
 import com.seafile.seadroid2.SeafException;
+import com.seafile.seadroid2.transfer.DownloadTaskInfo;
 import com.seafile.seadroid2.transfer.TransferService;
-import com.seafile.seadroid2.transfer.TransferManager.DownloadTaskInfo;
 import com.seafile.seadroid2.Utils;
 
 /**
